+++ Source/Core/FileSystem/L1Portability/BasicTCPSocket.h
@@ -123,7 +123,7 @@ public:
      * @brief Unsupported feature.
      * @return false.
      */
-    virtual bool RelativeSeek(const int32 deltaPos);
+    virtual bool RelativeSeek(const int64 deltaPos);
 
     /**
      * @brief Unsupported feature.
